fixing build
git-svn-id: http://svn.codehaus.org/geoserver/trunk@13812 ef1d6e69-97e5-0310-af46-8a06194da32a

Former-commit-id: 293d9b1e1d7eed85da114ca74075932a89843878 [formerly 60fe684fd48235b993c1edb07c0849a827751179]
Former-commit-id: da6dd348923e8ac03e0da80079bf4ff064a94864

+++ src/wms/src/test/java/org/geoserver/wms/kvp/GetMapKvpRequestReaderTest.java
@@ -132,7 +132,7 @@ public class GetMapKvpRequestReaderTest extends KvpRequestReaderTestSupport {
         assertEquals(1, request.getBuffer());
 
         assertEquals(PaletteManager.safePalette, request.getPalette().getIcm());
-        assertEquals(Integer.valueOf(4), request.getElevation());
+        assertEquals(Double.valueOf(4), request.getElevation());
         
         Calendar cal = Calendar.getInstance();
         cal.set(2006, 1, 27, 22, 8, 12);
